Move new components to separate folders
